third_party/WebKit/LayoutTests/fast/block/block-with-inline-replaced-child-following-text-expected.txt

Test for Bugzilla bug: 99442 Regression r130057: Improper preferred width calculation when an inline replaced object, wrapped in an inline flow, follows some text.



For the following, the width of the containing inline-block box should be 94px [goverened by the min-width of the text (64px) plus the padding (30px)].

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE

PASS getWidth('div1') is '94px'
PASS getWidth('div2') is '94px'
PASS getWidth('div3') is '94px'
PASS successfullyParsed is true

TEST COMPLETE
